undo calculate logic
